Renamed a bunch of screens, working on c02 screen
